requirements.txt
langchain
langchain-community
langchain-text-splitters
langchain-huggingface
langchain-google-genai
chromadb
sentence-transformers
python-dotenv
pypdf

fastapi
uvicorn
python-dotenv
langchain
langchain-huggingface
langchain-google-genai
chromadb
sentence-transformers
pydantic
aiofiles
python-multipart